backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
bcrypt==4.1.2
pyjwt==2.8.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==3.0.1

